Separate MlInformation and MlInformationList from Service
