+++ src/Tizen.MachineLearning.Inference/Tizen.MachineLearning.Inference/Service.cs
@@ -343,62 +343,4 @@ namespace Tizen.MachineLearning.Inference
             return new MlInformationList(infoList);
         }
     }
-
-    public class MlInformation {
-        private IntPtr _handle = IntPtr.Zero;
-
-        internal MlInformation(IntPtr handle) {
-            NNStreamer.CheckNNStreamerSupport();
-            _handle = handle;
-        }
-
-        ~MlInformation() {
-            if (_handle != IntPtr.Zero)
-            {
-                NNStreamerError ret = Interop.Util.DestroyInformation(_handle);
-                NNStreamer.CheckException(ret, "Failed to destroy the information");
-            }
-        }
-
-        public string GetInformation(string key)
-        {
-            IntPtr value;
-
-            NNStreamerError ret = Interop.Util.GetValue(_handle, key, out value);
-            NNStreamer.CheckException(ret, "Failed to get information value from key");
-            return Interop.Util.IntPtrToString(value);
-        }
-    }
-
-    public class MlInformationList {
-        private IntPtr _handle = IntPtr.Zero;
-
-        internal MlInformationList(IntPtr handle) {
-            NNStreamer.CheckNNStreamerSupport();
-            _handle = handle;
-        }
-
-        ~MlInformationList() {
-            NNStreamerError ret = Interop.Util.DestroyInformationList(_handle);
-            NNStreamer.CheckException(ret, "Failed to destroy the information list");
-        }
-
-        public MlInformation GetInformation(int index)
-        {
-            IntPtr value;
-
-            NNStreamerError ret = Interop.Util.GetInformation(_handle, index, out value);
-            NNStreamer.CheckException(ret, "Failed to get information from list");
-            return new MlInformation(value);
-        }
-
-        public int GetLength()
-        {
-            int value = 0;
-
-            NNStreamerError ret = Interop.Util.GetInformationListLength(_handle, out value);
-            NNStreamer.CheckException(ret, "Failed to get the length of information list");
-            return value;
-        }
-    }
 }
